+++ src/Controller/TopicsController.php
@@ -6,6 +6,7 @@ use App\Entity\Topics;
 use App\Form\TopicsType;
 use App\Repository\CategoryRepository;
 use App\Repository\ExposureRepository;
+use App\Repository\OpeningRepository;
 use App\Repository\TopicsRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -13,8 +14,9 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 
-#[Route('/topics')]
+#[Route('/sujets')]
 class TopicsController extends AbstractController
 {
     #[Route('/', name: 'app_topics_index', methods: ['GET', 'POST'])]
@@ -22,6 +24,7 @@ class TopicsController extends AbstractController
         TopicsRepository $topicsRepository,
         CategoryRepository $categoryRepository,
         ExposureRepository $exposureRepository,
+        OpeningRepository $openingRepository,
         Request $request
     ): Response {
 
@@ -40,6 +43,13 @@ class TopicsController extends AbstractController
         $categories = $categoryRepository->findAll();
         $exposures = $exposureRepository->findAll();
 
+        $openingHours = $openingRepository->findOneBy(['openingDay' => 'Lundi']);
+        $openingHourMorning = $openingHours->getOpeninghourmorning();
+        $closingHourMorning = $openingHours->getClosinghourmorning();
+        $openingHourAfternoon = $openingHours->getOpeninghourafternoon();
+        $closingHourAfternoon = $openingHours->getClosinghourafternoon();
+
+
         if ($request->get('ajax')) {
             if (!$topics) {
                 return new JsonResponse([
@@ -57,8 +67,6 @@ class TopicsController extends AbstractController
                     ]),
                 ]);
             }
-
-            
         }
 
         return $this->render('topics/index.html.twig', [
@@ -68,12 +76,21 @@ class TopicsController extends AbstractController
             'page' => $page,
             'categories' => $categories,
             'exposures' => $exposures,
+            'opening' => $openingRepository->findAll(),
+            'openingHourMorning' => $openingHourMorning,
+            'closingHourMorning' => $closingHourMorning,
+            'openingHourAfternoon' => $openingHourAfternoon,
+            'closingHourAfternoon' => $closingHourAfternoon,
         ]);
     }
 
     #[Route('/new', name: 'app_topics_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé.');
+        }
+
         $topic = new Topics();
         $form = $this->createForm(TopicsType::class, $topic);
         $form->handleRequest($request);
@@ -92,16 +109,31 @@ class TopicsController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_topics_show', methods: ['GET'])]
-    public function show(Topics $topic): Response
+    public function show(Topics $topic, OpeningRepository $openingRepository): Response
     {
+        $openingHours = $openingRepository->findOneBy(['openingDay' => 'Lundi']);
+        $openingHourMorning = $openingHours->getOpeninghourmorning();
+        $closingHourMorning = $openingHours->getClosinghourmorning();
+        $openingHourAfternoon = $openingHours->getOpeninghourafternoon();
+        $closingHourAfternoon = $openingHours->getClosinghourafternoon();
+        
         return $this->render('topics/show.html.twig', [
             'topic' => $topic,
+            'opening' => $openingRepository->findAll(),
+            'openingHourMorning' => $openingHourMorning,
+            'closingHourMorning' => $closingHourMorning,
+            'openingHourAfternoon' => $openingHourAfternoon,
+            'closingHourAfternoon' => $closingHourAfternoon,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_topics_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Topics $topic, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé.');
+        }
+
         $form = $this->createForm(TopicsType::class, $topic);
         $form->handleRequest($request);
 
@@ -120,6 +152,10 @@ class TopicsController extends AbstractController
     #[Route('/{id}', name: 'app_topics_delete', methods: ['POST'])]
     public function delete(Request $request, Topics $topic, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            throw new AccessDeniedException('Accès refusé.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$topic->getId(), $request->request->get('_token'))) {
             $entityManager->remove($topic);
             $entityManager->flush();
